requirements.txt
# Flask and extensions
flask==2.3.3
flask-wtf==1.2.1
flask-login==0.6.3
python-dotenv==1.0.0

# API client
requests==2.31.0
httpx==0.24.1

# Data processing
pandas==2.0.3
numpy==1.24.4

# Visualization
plotly==5.17.0
dash==2.13.0

# Utilities
python-dateutil==2.8.2

# Development tools
pytest==7.4.0
black==23.7.0